requirements.txt
pyopenssl
pyyaml